# swaggerDocs/swagger.yaml
basePath: /kapis/kubeeye.kubesphere.io/v1alpha1
definitions:
...
  v1alpha2.ClusterInfo:
    properties:
      namespacesCount:
        type: integer
      nodesCount:
        type: integer
      version:
        type: string
    type: object
  v1alpha2.ComponentResultItem:
    properties:
      endpoint:
        type: string
      name:
        type: string
      namespace:
        type: string
    type: object
  v1alpha2.ExtraInfo:
    properties:
      namespacesList:
        items:
          type: string
        type: array
      workloadsCount:
        type: integer
    type: object
  v1alpha2.FileChangeResultItem:
    properties:
      fileName:
        type: string
      issues:
        items:
          type: string
        type: array
      path:
        type: string
    type: object
  v1alpha2.FileChangeRule:
    properties:
      desc:
        type: string
      level:
        $ref: '#/definitions/v1alpha2.Level'
      name:
        type: string
      nodeName:
        type: string
      nodeSelector:
        additionalProperties:
          type: string
        type: object
      path:
        type: string
      rule:
        type: string
    type: object
  v1alpha2.FileFilterRule:
    properties:
      desc:
        type: string
      level:
        $ref: '#/definitions/v1alpha2.Level'
      name:
        type: string
      nodeName:
        type: string
      nodeSelector:
        additionalProperties:
          type: string
        type: object
      path:
        type: string
      rule:
        type: string
    type: object
  v1alpha2.InspectPlan:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1alpha2.InspectPlanSpec'
      status:
        $ref: '#/definitions/v1alpha2.InspectPlanStatus'
    type: object
  v1alpha2.InspectPlanSpec:
    properties:
      clusterName:
        items:
          type: string
        type: array
      kubeConfig:
        type: string
      maxTasks:
        type: integer
      ruleNames:
        items:
          type: string
        type: array
      schedule:
        type: string
      suspend:
        type: boolean
      tag:
        type: string
      timeout:
        type: string
    type: object
  v1alpha2.InspectPlanStatus:
    properties:
      TaskNames:
        items:
          $ref: '#/definitions/v1alpha2.TaskStatus'
        type: array
      lastScheduleTime:
        type: string
      lastTaskName:
        type: string
      lastTaskStatus:
        $ref: '#/definitions/v1alpha2.Phase'
      nextScheduleTime:
        type: string
    type: object
  v1alpha2.InspectResult:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1alpha2.InspectResultSpec'
      status:
        $ref: '#/definitions/v1alpha2.InspectResultStatus'
    type: object
  v1alpha2.InspectResultSpec:
    properties:
      componentResult:
        items:
          $ref: '#/definitions/v1alpha2.ComponentResultItem'
        type: array
      inspectRuleTotal:
        additionalProperties:
          type: integer
        description: |-
          INSERT ADDITIONAL SPEC FIELDS - desired state of cluster
          Important: Run "make" to regenerate code after modifying this file
        type: object
      nodeInfoResult:
        additionalProperties:
          $ref: '#/definitions/v1alpha2.NodeInfoResult'
        type: object
      opaResult:
        $ref: '#/definitions/v1alpha2.KubeeyeOpaResult'
      prometheusResult:
        items:
          items:
            additionalProperties:
              type: string
            type: object
          type: array
        type: array
    type: object
  v1alpha2.InspectResultStatus:
    properties:
      complete:
        type: boolean
      duration:
        type: string
      level:
        additionalProperties:
          type: integer
        type: object
      policy:
        $ref: '#/definitions/v1alpha2.Policy'
      taskEndTime:
        type: string
      taskStartTime:
        type: string
    type: object
  v1alpha2.InspectRule:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1alpha2.InspectRuleSpec'
      status:
        $ref: '#/definitions/v1alpha2.InspectRuleStatus'
    type: object
  v1alpha2.InspectRuleSpec:
    properties:
      component:
        type: string
      fileChange:
        items:
          $ref: '#/definitions/v1alpha2.FileChangeRule'
        type: array
      fileFilter:
        items:
          $ref: '#/definitions/v1alpha2.FileFilterRule'
        type: array
      opas:
        items:
          $ref: '#/definitions/v1alpha2.OpaRule'
        type: array
      prometheus:
        items:
          $ref: '#/definitions/v1alpha2.PrometheusRule'
        type: array
      prometheusEndpoint:
        type: string
      sysctl:
        items:
          $ref: '#/definitions/v1alpha2.SysRule'
        type: array
      systemd:
        items:
          $ref: '#/definitions/v1alpha2.SysRule'
        type: array
    type: object
  v1alpha2.InspectRuleStatus:
    properties:
      endImportTime:
        type: string
      ruleCount:
        type: integer
      startImportTime:
        type: string
      state:
        $ref: '#/definitions/v1alpha2.State'
    type: object
  v1alpha2.InspectTask:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1alpha2.InspectTaskSpec'
      status:
        $ref: '#/definitions/v1alpha2.InspectTaskStatus'
    type: object
  v1alpha2.InspectTaskSpec:
    properties:
      clusterName:
        items:
          type: string
        type: array
      inspectPolicy:
        $ref: '#/definitions/v1alpha2.Policy'
      timeout:
        type: string
    type: object
  v1alpha2.InspectTaskStatus:
    properties:
      clusterInfo:
        allOf:
        - $ref: '#/definitions/v1alpha2.ClusterInfo'
        description: |-
          INSERT ADDITIONAL STATUS FIELD - define observed state of cluster
          Important: Run "make" to regenerate code after modifying this file
      endTimestamp:
        type: string
      phase:
        items:
          $ref: '#/definitions/v1alpha2.JobPhase'
        type: array
      startTimestamp:
        type: string
    type: object
  v1alpha2.JobPhase:
    properties:
      jobName:
        type: string
      phase:
        $ref: '#/definitions/v1alpha2.Phase'
    type: object
  v1alpha2.KubeeyeOpaResult:
    properties:
      extraInfo:
        $ref: '#/definitions/v1alpha2.ExtraInfo'
      percent:
        type: integer
      resourceResults:
        items:
          $ref: '#/definitions/v1alpha2.ResourceResult'
        type: array
      scoreInfo:
        $ref: '#/definitions/v1alpha2.ScoreInfo'
    type: object
  v1alpha2.Level:
    enum:
    - danger
    - warning
    - ignore
    type: string
    x-enum-varnames:
    - DangerLevel
    - WarningLevel
    - IgnoreLevel
  v1alpha2.NodeInfoResult:
    properties:
      fileChangeResult:
        items:
          $ref: '#/definitions/v1alpha2.FileChangeResultItem'
        type: array
      fileFilterResult:
        items:
          $ref: '#/definitions/v1alpha2.FileChangeResultItem'
        type: array
      nodeInfo:
        additionalProperties:
          type: string
        type: object
      sysctlResult:
        items:
          $ref: '#/definitions/v1alpha2.NodeResultItem'
        type: array
      systemdResult:
        items:
          $ref: '#/definitions/v1alpha2.NodeResultItem'
        type: array
    type: object
  v1alpha2.NodeResultItem:
    properties:
      assert:
        type: boolean
      name:
        type: string
      value:
        type: string
    type: object
  v1alpha2.OpaRule:
    properties:
      desc:
        type: string
      level:
        $ref: '#/definitions/v1alpha2.Level'
      module:
        type: string
      name:
        type: string
      rule:
        type: string
    type: object
  v1alpha2.Phase:
    enum:
    - Pending
    - Running
    - Succeeded
    - Failed
    - Unknown
    type: string
    x-enum-varnames:
    - PhasePending
    - PhaseRunning
    - PhaseSucceeded
    - PhaseFailed
    - PhaseUnknown
  v1alpha2.Policy:
    enum:
    - timing
    - instant
    type: string
    x-enum-varnames:
    - TimingPolicy
    - InstantPolicy
  v1alpha2.PrometheusRule:
    properties:
      desc:
        type: string
      endpoint:
        type: string
      level:
        $ref: '#/definitions/v1alpha2.Level'
      name:
        type: string
      rule:
        type: string
    type: object
  v1alpha2.ResourceResult:
    properties:
      name:
        type: string
      namespace:
        type: string
      resourceType:
        type: string
      resultItems:
        items:
          $ref: '#/definitions/v1alpha2.ResultItem'
        type: array
    type: object
  v1alpha2.ResultItem:
    properties:
      level:
        type: string
      message:
        type: string
      reason:
        type: string
    type: object
  v1alpha2.ScoreInfo:
    properties:
      dangerous:
        type: integer
      ignore:
        type: integer
      passing:
        type: integer
      score:
        type: integer
      total:
        type: integer
      warning:
        type: integer
    type: object
  v1alpha2.State:
    enum:
    - StartImport
    - ImportComplete
    type: string
    x-enum-varnames:
    - StartImport
    - ImportComplete
  v1alpha2.SysRule:
    properties:
      desc:
        type: string
      level:
        $ref: '#/definitions/v1alpha2.Level'
      name:
        type: string
      nodeName:
        type: string
      nodeSelector:
        additionalProperties:
          type: string
        type: object
      rule:
        type: string
    type: object
  v1alpha2.TaskStatus:
    properties:
      name:
        type: string
      taskStatus:
        $ref: '#/definitions/v1alpha2.Phase'
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a kubeeye api server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: KubeEye API
  version: "1.0"
paths:
  /inspectplans:
    delete:
      consumes:
      - application/json
      description: DeleteInspectTask
      parameters:
      - description: delete InspectTask
        in: body
        name: v1alpha2.InspectTask
        required: true
        schema:
          $ref: '#/definitions/v1alpha2.InspectTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha2.InspectTask'
      summary: Show an Inspect
      tags:
      - InspectTask
    get:
      consumes:
      - application/json
      description: ListInspectPlan
      parameters:
      - description: orderBy
        in: query
        name: orderBy
        type: string
      - description: ascending
        in: query
        name: ascending
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1alpha2.InspectPlan'
            type: array
      summary: Show an Inspect
      tags:
      - InspectPlan
    post:
      consumes:
      - application/json
      description: CreateInspectPlan
      parameters:
      - description: Add InspectPlan
        in: body
        name: InspectPlan
        required: true
        schema:
          $ref: '#/definitions/v1alpha2.InspectPlan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha2.InspectPlan'
      summary: Show an Inspect
      tags:
      - InspectPlan
  /inspectplans/{name}:
    get:
      consumes:
      - application/json
      description: GetInspectPlan
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha2.InspectPlan'
      summary: Show an Inspect
      tags:
      - InspectPlan
  /inspectresults:
    get:
      consumes:
      - application/json
      description: get ListInspectResult
      parameters:
      - description: orderBy
        in: query
        name: orderBy
        type: string
      - description: ascending
        in: query
        name: ascending
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1alpha2.InspectResult'
            type: array
      summary: Show an Inspect
      tags:
      - InspectResult
  /inspectresults/{name}:
    get:
      consumes:
      - application/json
      description: GetInspectResult
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha2.InspectResult'
      summary: Show an Inspect
      tags:
      - InspectResult
  /inspectrules:
    delete:
      consumes:
      - application/json
      description: DeleteInspectRule
      parameters:
      - description: Add InspectRule
        in: body
        name: v1alpha2.InspectRule
        required: true
        schema:
          $ref: '#/definitions/v1alpha2.InspectRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha2.InspectRule'
      summary: Show an Inspect
      tags:
      - InspectRule
    get:
      consumes:
      - application/json
      description: ListInspectRule
      parameters:
      - description: orderBy
        in: query
        name: orderBy
        type: string
      - description: ascending
        in: query
        name: ascending
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1alpha2.InspectRule'
            type: array
      summary: Show an Inspect
      tags:
      - InspectRule
    post:
      consumes:
      - application/json
      description: CreateInspectRule
      parameters:
      - description: Add InspectRule
        in: body
        name: InspectRule
        required: true
        schema:
          $ref: '#/definitions/v1alpha2.InspectRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha2.InspectRule'
      summary: Show an Inspect
      tags:
      - InspectRule
    put:
      consumes:
      - application/json
      description: UpdateInspectRule
      parameters:
      - description: Add InspectRule
        in: body
        name: v1alpha2.InspectRule
        required: true
        schema:
          $ref: '#/definitions/v1alpha2.InspectRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha2.InspectRule'
      summary: Show an Inspect
      tags:
      - InspectRule
  /inspectrules/{name}:
    get:
      consumes:
      - application/json
      description: GetInspectRule
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha2.InspectRule'
      summary: Show an Inspect
      tags:
      - InspectRule
  /inspecttasks:
    get:
      consumes:
      - application/json
      description: ListInspectTask
      parameters:
      - description: orderBy
        in: query
        name: orderBy
        type: string
      - description: ascending
        in: query
        name: ascending
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1alpha2.InspectTask'
            type: array
      summary: Show an Inspect
      tags:
      - InspectTask
  /inspecttasks/{name}:
    get:
      consumes:
      - application/json
      description: ListInspectTask
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha2.InspectTask'
      summary: Show an Inspect
      tags:
      - InspectTask
swagger: "2.0"